customuser/serializers.py
from rest_framework import serializers
from customuser.models import CustomUser


class UserSerializer(serializers.ModelSerializer):

    class Meta:
        model = CustomUser
        fields = ['id', 'name', 'username', 'email']
